+++ crates/ruma-client-api/src/rendezvous.rs
@@ -6,6 +6,4 @@ pub mod delete_rendezvous_session;
 #[cfg(feature = "unstable-msc4388")]
 pub mod get_rendezvous_session;
 #[cfg(feature = "unstable-msc4388")]
-pub mod unstable_create_rendezvous_session;
-#[cfg(feature = "unstable-msc4388")]
 pub mod update_rendezvous_session;
