+++ crates/ruma-client-api/src/rendezvous/update_rendezvous_session.rs
@@ -3,7 +3,7 @@
 //! Update a rendezvous session.
 
 pub mod unstable {
-    //! `msc4388` ([MSC])
+    //! `unstable/io.element.msc4388` ([MSC])
     //!
     //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4388
